Fix for issue #55: ThreadMonitor waits unnecessarily
This change provides a fix and test for issue #55:
https://github.com/javadelight/delight-nashorn-sandbox/issues/55

In ThreadMonitor in the call to wait on the monitor object. It's
possible that the threadToMonitor was already set and the monitor
notification sent. In that case, the monitor wait goes to the
maximum value for the cpu time limit.

The fix is to first verify that the threadToMonitor is null within
the synchronized block and before the wait on the monitor.