Make it work on JDK 8, 11 and 15
Due to the use of both JDK (removed from 15) and Standalone Nashorn (compiled to 11), we must build on JDK 11 with source and target 1.8.
Then we can run the tests on JDK 8 and 15 with maven verify.
We must not clean or change the sources in between the build on JDK 11 and the tests on other JDKs otherwise, Maven will rebuild the module and will fail.